+++ src/Bitcoin/TransactionBuilder.cpp
@@ -104,7 +104,7 @@ TransactionPlan TransactionBuilder::plan(const SigningInput& input) {
 
         // if amount requested is the same or more than available amount, it cannot be satisifed, but
         // treat this case as MaxAmount, and send maximum available (which will be less)
-        if (!maxAmount && input.amount >= inputSum) {
+        if (!maxAmount && static_cast<uint64_t>(input.amount) >= inputSum) {
             maxAmount = true;
         }
 
@@ -127,7 +127,7 @@ TransactionPlan TransactionBuilder::plan(const SigningInput& input) {
         } else {
             // truncate to limit number of selected UTXOs
             plan.utxos.clear();
-            for (auto i = 0; i < MaxUtxosHardLimit; ++i) {
+            for (auto i = 0ul; i < MaxUtxosHardLimit; ++i) {
                 plan.utxos.push_back(selectedInputs[i]);
             }
         }
